web/transiq/requirements.txt
amqp==2.2.2
Babel==2.5.3
beautifulsoup4==4.6.0
billiard==3.5.0.3
boto3==1.7.10
botocore==1.10.10
bs4==0.0.1
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
Django==2.1.4
django-filter==2.0.0
django-simple-history==2.0
django-storages==1.6.6
djangorestframework==3.8.2
django-admin-rangefilter==0.3.6
docutils==0.14
googletrans==2.2.0
idna==2.6
jmespath==0.9.3
jsonschema==2.5.1
kombu==4.1.0
lxml==4.2.1
matplotlib==2.2.3
model-mommy==1.5.1
num2words==0.5.6
numpy==1.14.3
oauthlib==2.0.7
openpyxl==2.5.3
pandas==0.22.0
pdfkit==0.6.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pygeocoder==1.2.5
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.13
shortuuid==0.5.0
six==1.11.0
slugify==0.0.1
unittest-xml-reporting==2.1.0
urllib3==1.22
vine==1.1.4
xlrd==1.1.0
arrow==0.12.1
coverage==4.5.1
selenium==3.11.0
python-slugify==1.2.5
fcm-django==0.2.21
djangorestframework-datatables==0.4.0
pynamodb==3.3.1
